+++ graphdoc/graphdoc/data/dspy_data/__init__.py
@@ -1,9 +1,9 @@
 # Copyright 2025-, Semiotic AI, Inc.
 # SPDX-License-Identifier: Apache-2.0
 
-from .dspy_data_helper import DspyDataHelper
-from .generation_data_helper import GenerationDataHelper
-from .quality_data_helper import QualityDataHelper
+from graphdoc.data.dspy_data.dspy_data_helper import DspyDataHelper
+from graphdoc.data.dspy_data.generation_data_helper import GenerationDataHelper
+from graphdoc.data.dspy_data.quality_data_helper import QualityDataHelper
 
 __all__ = [
     "DspyDataHelper",
