+++ graphdoc/graphdoc/data/mlflow_data/__init__.py
@@ -1,7 +1,7 @@
 # Copyright 2025-, Semiotic AI, Inc.
 # SPDX-License-Identifier: Apache-2.0
 
-from .mlflow_data_helper import MlflowDataHelper
+from graphdoc.data.mlflow_data.mlflow_data_helper import MlflowDataHelper
 
 __all__ = [
     "MlflowDataHelper",
