+++ graphdoc/graphdoc/eval/__init__.py
@@ -1,7 +1,7 @@
 # Copyright 2025-, Semiotic AI, Inc.
 # SPDX-License-Identifier: Apache-2.0
 
-from .doc_generator_eval import DocGeneratorEvaluator
+from graphdoc.eval.doc_generator_eval import DocGeneratorEvaluator
 
 __all__ = [
     "DocGeneratorEvaluator",
