+++ graphdoc/graphdoc/modules/__init__.py
@@ -1,7 +1,7 @@
 # Copyright 2025-, Semiotic AI, Inc.
 # SPDX-License-Identifier: Apache-2.0
 
-from .doc_generator_module import DocGeneratorModule
+from graphdoc.modules.doc_generator_module import DocGeneratorModule
 
 __all__ = [
     "DocGeneratorModule",
