+++ graphdoc/graphdoc/prompts/__init__.py
@@ -8,20 +8,20 @@ from typing import Union
 # external packages
 import dspy
 
-from .schema_doc_generation import (
+from graphdoc.prompts.schema_doc_generation import (
     BadDocGeneratorSignature,
     DocGeneratorHelperSignature,
     DocGeneratorPrompt,
     DocGeneratorSignature,
 )
-from .schema_doc_quality import (
+from graphdoc.prompts.schema_doc_quality import (
     DocQualityDemonstrationSignature,
     DocQualityPrompt,
     DocQualitySignature,
 )
 
 # internal packages
-from .single_prompt import SinglePrompt
+from graphdoc.prompts.single_prompt import SinglePrompt
 
 # logging
 log = logging.getLogger(__name__)
