+++ graphdoc/graphdoc/train/__init__.py
@@ -8,12 +8,15 @@ from typing import Any, Dict, List
 # external packages
 import dspy
 
-from .doc_generator_trainer import DocGeneratorTrainer
-from .doc_quality_trainer import DocQualityTrainer
-
 # internal packages
-from .optimizers import _optimizer_kwargs_filter, optimizer_class, optimizer_compile
-from .single_prompt_trainer import SinglePrompt, SinglePromptTrainer
+from graphdoc.train.doc_generator_trainer import DocGeneratorTrainer
+from graphdoc.train.doc_quality_trainer import DocQualityTrainer
+from graphdoc.train.optimizers import (
+    _optimizer_kwargs_filter,
+    optimizer_class,
+    optimizer_compile,
+)
+from graphdoc.train.single_prompt_trainer import SinglePrompt, SinglePromptTrainer
 
 # logging
 log = logging.getLogger(__name__)
